refactor (webapi): move sipi responders to store package
